fix: mount linked worktree git dirs in sandbox (#127)
