+++ src/sandbox.rs
@@ -4,6 +4,7 @@
 use std::process::{Command, Stdio};
 
 use crate::types::AgentKind;
+use crate::worktree_layout::{read_commondir, resolve_worktree_gitdir};
 
 const CONTAINER_BIN: &str = "container";
 const SANDBOX_IMAGE: &str = "aid-sandbox:latest";
@@ -55,11 +56,28 @@ pub fn wrap_command(
         wrapped.arg("--tmpfs").arg("/tmp");
     }
     if let Some(dir) = cwd.as_deref() {
-        wrapped.arg("-v").arg(format!("{dir}:{dir}"));
-        wrapped.arg("-w").arg(dir);
-        wrapped.current_dir(dir);
+        mount_workdir(&mut wrapped, dir);
     }
-    // Forward env vars explicitly set on the command
+    forward_command_envs(&mut wrapped, cmd);
+    forward_agent_envs(&mut wrapped, agent_kind);
+    mount_agent_home(&mut wrapped, agent_kind);
+    wrapped.arg("-e").arg("HOME=/root");
+    wrapped.arg(SANDBOX_IMAGE);
+    wrapped.arg(cmd.get_program());
+    wrapped.args(cmd.get_args());
+    wrapped
+}
+
+fn mount_workdir(wrapped: &mut Command, dir: &str) {
+    wrapped.arg("-v").arg(format!("{dir}:{dir}"));
+    for (host, container) in worktree_git_mounts(dir) {
+        wrapped.arg("-v").arg(format!("{host}:{container}"));
+    }
+    wrapped.arg("-w").arg(dir);
+    wrapped.current_dir(dir);
+}
+
+fn forward_command_envs(wrapped: &mut Command, cmd: &Command) {
     for (key, value) in cmd.get_envs() {
         if let Some(value) = value {
             wrapped.arg("-e").arg(format!(
@@ -69,13 +87,17 @@ pub fn wrap_command(
             ));
         }
     }
-    // Forward agent-specific API keys from host environment (inherit mode)
+}
+
+fn forward_agent_envs(wrapped: &mut Command, agent_kind: AgentKind) {
     for key in agent_env_keys(agent_kind) {
         if std::env::var_os(key).is_some() {
             wrapped.arg("-e").arg(*key);
         }
     }
-    // Mount agent config directories (auth, settings) from host
+}
+
+fn mount_agent_home(wrapped: &mut Command, agent_kind: AgentKind) {
     if let Some(home) = std::env::var_os("HOME") {
         let home = std::path::Path::new(&home);
         for subdir in agent_config_dirs(agent_kind) {
@@ -98,11 +120,37 @@ pub fn wrap_command(
                 .arg("AID_HOME=/root/.aid");
         }
     }
-    wrapped.arg("-e").arg("HOME=/root");
-    wrapped.arg(SANDBOX_IMAGE);
-    wrapped.arg(cmd.get_program());
-    wrapped.args(cmd.get_args());
-    wrapped
+}
+
+fn worktree_git_mounts(cwd: &str) -> Vec<(String, String)> {
+    let cwd_path = std::path::Path::new(cwd);
+    let cwd_mount = std::fs::canonicalize(cwd_path).unwrap_or_else(|_| cwd_path.to_path_buf());
+    let Some(gitdir) = resolve_worktree_gitdir(cwd_path) else {
+        return Vec::new();
+    };
+    let mut mounts = Vec::new();
+    push_git_mount(&mut mounts, &cwd_mount, &gitdir);
+    if let Some(commondir) = read_commondir(&gitdir) {
+        push_git_mount(&mut mounts, &cwd_mount, &commondir);
+    }
+    mounts
+}
+
+fn push_git_mount(
+    mounts: &mut Vec<(String, String)>,
+    cwd: &std::path::Path,
+    path: &std::path::Path,
+) {
+    if path.starts_with(cwd) {
+        return;
+    }
+    let Some(mount) = path.to_str().map(ToOwned::to_owned) else {
+        return;
+    };
+    if mounts.iter().any(|(host, _)| host == &mount) {
+        return;
+    }
+    mounts.push((mount.clone(), mount));
 }
 
 fn agent_env_keys(kind: AgentKind) -> &'static [&'static str] {
@@ -144,154 +192,4 @@ pub fn kill_container(task_id: &str) {
         .stdout(Stdio::null())
         .stderr(Stdio::null())
         .status();
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{can_sandbox, wrap_command};
-    use crate::types::AgentKind;
-    use std::{
-        ffi::OsString,
-        fs,
-        process::Command,
-        sync::{Mutex, OnceLock},
-    };
-    use tempfile::tempdir;
-
-    fn args(cmd: &Command) -> Vec<String> {
-        cmd.get_args()
-            .map(|arg| arg.to_string_lossy().into_owned())
-            .collect()
-    }
-
-    fn env_lock() -> &'static Mutex<()> {
-        static LOCK: OnceLock<Mutex<()>> = OnceLock::new();
-        LOCK.get_or_init(|| Mutex::new(()))
-    }
-
-    struct HomeGuard(Option<OsString>);
-
-    impl Drop for HomeGuard {
-        fn drop(&mut self) {
-            match self.0.take() {
-                Some(home) => unsafe {
-                    std::env::set_var("HOME", home);
-                },
-                None => unsafe {
-                    std::env::remove_var("HOME");
-                },
-            }
-        }
-    }
-
-    fn with_home<F>(dirs: &[&str], test: F)
-    where
-        F: FnOnce(),
-    {
-        let _guard = env_lock().lock().expect("env lock poisoned");
-        let temp = tempdir().expect("tempdir");
-        for dir in dirs {
-            fs::create_dir_all(temp.path().join(dir)).expect("create home subdir");
-        }
-        let original_home = std::env::var_os("HOME");
-        let _home_guard = HomeGuard(original_home);
-        unsafe {
-            std::env::set_var("HOME", temp.path());
-        }
-        test();
-    }
-
-    #[test]
-    fn cannot_sandbox_native_agents() {
-        assert!(!can_sandbox(AgentKind::OpenCode));
-        assert!(!can_sandbox(AgentKind::Copilot));
-        assert!(!can_sandbox(AgentKind::Cursor));
-        assert!(!can_sandbox(AgentKind::Droid));
-        assert!(!can_sandbox(AgentKind::Oz));
-        assert!(!can_sandbox(AgentKind::Claude));
-        assert!(!can_sandbox(AgentKind::Custom));
-        assert!(can_sandbox(AgentKind::Codex));
-    }
-
-    #[test]
-    fn wrap_command_builds_container_run() {
-        let mut cmd = Command::new("codex");
-        cmd.args(["exec", "ship it"]);
-
-        let wrapped = wrap_command(&cmd, "t-abcd", AgentKind::Codex, false);
-        let wrapped_args = args(&wrapped);
-
-        assert_eq!(wrapped.get_program().to_string_lossy(), "container");
-        assert!(wrapped_args.iter().any(|arg| arg == "run"));
-        assert!(wrapped_args.iter().any(|arg| arg == "--rm"));
-        assert!(wrapped_args.iter().any(|arg| arg == "--init"));
-        assert!(wrapped_args.iter().any(|arg| arg == "aid-sandbox:latest"));
-        assert_eq!(wrapped_args[wrapped_args.len() - 3], "codex");
-        assert_eq!(wrapped_args[wrapped_args.len() - 2], "exec");
-        assert_eq!(wrapped_args[wrapped_args.len() - 1], "ship it");
-    }
-
-    #[test]
-    fn wrap_command_forwards_env_vars() {
-        let mut cmd = Command::new("codex");
-        cmd.env("OPENAI_API_KEY", "test-key");
-
-        let wrapped = wrap_command(&cmd, "t-abcd", AgentKind::Codex, false);
-        let wrapped_args = args(&wrapped);
-
-        assert!(wrapped_args
-            .windows(2)
-            .any(|pair| pair == ["-e", "OPENAI_API_KEY=test-key"]));
-    }
-
-    #[test]
-    fn wrap_command_mounts_project_dir() {
-        with_home(&[".aid"], || {
-            let mut cmd = Command::new("codex");
-            cmd.current_dir("/tmp/project");
-
-            let wrapped = wrap_command(&cmd, "t-abcd", AgentKind::Codex, false);
-            let wrapped_args = args(&wrapped);
-
-            assert!(wrapped_args
-                .windows(2)
-                .any(|pair| pair == ["-v", "/tmp/project:/tmp/project"]));
-            assert!(wrapped_args
-                .windows(2)
-                .any(|pair| pair == ["-w", "/tmp/project"]));
-            assert!(wrapped_args
-                .windows(2)
-                .any(|pair| pair[0] == "-v" && pair[1].ends_with(":/root/.aid")));
-        });
-    }
-
-    #[test]
-    fn wrap_command_mounts_aid_home() {
-        with_home(&[".aid"], || {
-            let cmd = Command::new("codex");
-
-            let wrapped = wrap_command(&cmd, "t-abcd", AgentKind::Codex, false);
-            let wrapped_args = args(&wrapped);
-
-            assert!(wrapped_args
-                .windows(2)
-                .any(|pair| pair[0] == "-v" && pair[1].ends_with(":/root/.aid")));
-            assert!(wrapped_args
-                .windows(2)
-                .any(|pair| pair == ["-e", "AID_HOME=/root/.aid"]));
-        });
-    }
-
-    #[test]
-    fn wrap_command_readonly_adds_flag() {
-        let cmd = Command::new("codex");
-
-        let wrapped = wrap_command(&cmd, "t-abcd", AgentKind::Codex, true);
-        let wrapped_args = args(&wrapped);
-
-        assert!(wrapped_args.iter().any(|arg| arg == "--read-only"));
-        assert!(wrapped_args
-            .windows(2)
-            .any(|pair| pair == ["--tmpfs", "/tmp"]));
-    }
 }
